+++ newserver.h
@@ -11,7 +11,7 @@ std::pair<char*, int> errorMessage ();
 
 // функция принимает аргументами структуры RelayData, которые заполняются просто установкой полей структуры
 // возвращает char* выделенной памяти !НЕ ЗАБЫТЬ ПОЧИСТИТЬ!
-std::pair<char*, int> encodeSensorsData(RelayData solar, RelayData wind, RelayData generator, int batteryVoltage, RelayData consumers[RELAYS_NUMBER]);
+std::pair<char*, int> encodeSensorsData(RelayData solar, RelayData wind, RelayData generator, double batteryVoltage, RelayData consumers[RELAYS_NUMBER]);
 
 // функция принимает логин и пароль, возвращает также буфер, который необходимо отправить
 // !НЕ ЗАБЫТЬ ПОЧИСТИТЬ!
